requirements.txt
disnake
mysql-connector-python
asyncio
pyyaml
fastapi
uvicorn
